Refactor cryptographic stream handling and improve packet processing
- Updated `Crypt` implementation to use constant tag length for buffer capacity.
- Enhanced `decrypt` method to correctly handle lengths of encrypted data.
- Removed unused `client.rs` and `server.rs` stream implementations.
- Introduced generic stream handling for client and server with `TunnelClientInboundStream` and `TunnelServerInboundStream`.
- Added tests for inbound and outbound stream functionality, including error handling and multiple packet processing.
- Improved logging for better traceability during stream operations.

+++ crates/server/src/crypt/consts.rs
@@ -1,4 +1,4 @@
-pub const MAX_PACKET_SIZE: u16 = 4096; // Hard limit for packet size. Anythig abobe this will be rejected.
+pub const MAX_PACKET_SIZE: usize = 4096; // Hard limit for packet size. Anythig abobe this will be rejected.
 pub const HEADER_LENGTH: usize = 8 + 2; // counter (8 bytes) + length (2 bytes)
 pub const TAG_LENGTH: usize = 16; // AES-GCM tag length
 // IPv6 minimum MTU is 1280 bytes, minus IP (40 bytes) and UDP (8 bytes, future) headers - leaves 1232 bytes for payload
